Splitting out protocol flavors for integration tests.
