+++ test/protocols_flavor.py
@@ -46,81 +46,7 @@ class ProtocolsFlavor(object):
     raise NotImplementedError('Not implemented in the base class')
 
 
-class GoRpcProtocolsFlavor(ProtocolsFlavor):
-  """Overrides to use go rpc everywhere"""
-
-  def binlog_player_protocol(self):
-    return 'gorpc'
-
-  def vtctl_client_protocol(self):
-    return 'gorpc'
-
-  def vtctl_python_client_protocol(self):
-    return 'gorpc'
-
-  def vtworker_client_protocol(self):
-    # There is no GoRPC implementation for the vtworker RPC interface,
-    # so we use gRPC as well.
-    return 'grpc'
-
-  def tablet_manager_protocol(self):
-    return 'bson'
-
-  def tabletconn_protocol(self):
-    return 'gorpc'
-
-  def vtgate_protocol_flags(self):
-    return ['-vtgate_protocol', 'gorpc']
-
-  def rpc_timeout_message(self):
-    return 'timeout waiting for'
-
-  def service_map(self):
-    return ['grpc-vtworker']
-
-
-class GRpcProtocolsFlavor(ProtocolsFlavor):
-  """Overrides to use gRPC everywhere where it is supported.
-  If not supported yet, use GoRPC."""
-
-  def binlog_player_protocol(self):
-    return 'grpc'
-
-  def vtctl_client_protocol(self):
-    return 'grpc'
-
-  def vtctl_python_client_protocol(self):
-    return 'gorpc'
-
-  def vtworker_client_protocol(self):
-    return 'grpc'
-
-  def tablet_manager_protocol(self):
-    return 'grpc'
-
-  def tabletconn_protocol(self):
-    return 'grpc'
-
-  def vtgate_protocol_flags(self):
-    return ['-vtgate_protocol', 'gorpc']
-
-  def rpc_timeout_message(self):
-    return 'context deadline exceeded'
-
-  def service_map(self):
-    return [
-        'grpc-queryservice',
-        'grpc-updatestream',
-        'grpc-vtctl',
-        'grpc-vtworker',
-        'grpc-tabletmanager',
-        ]
-
-
-__knows_protocols_flavor_map = {
-  'gorpc': GoRpcProtocolsFlavor,
-  'grpc': GRpcProtocolsFlavor,
-}
+__knows_protocols_flavor_map = {}
 __protocols_flavor = None
 
 def protocols_flavor():
